third_party/WebKit/LayoutTests/imported/wpt/IndexedDB/idbobjectstore_delete-expected.txt

This is a testharness.js-based test.
FAIL IDBObjectStore.delete() - delete removes record (inline keys) assert_equals: expected (undefined) undefined but got (number) 1
Harness: the test ran to completion.
